Extract common tests to SpecialTypeKeywordRecommenderTests
