+++ src/EditorFeatures/CSharpTest2/Recommendations/KeywordRecommenderTests.cs
@@ -21,7 +21,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.Recommendations
         {
             foreach (var recommenderType in typeof(AbstractSyntacticSingleKeywordRecommender).Assembly.GetTypes())
             {
-                if (recommenderType.IsSubclassOf(typeof(AbstractSyntacticSingleKeywordRecommender)))
+                if (!recommenderType.IsAbstract && recommenderType.IsSubclassOf(typeof(AbstractSyntacticSingleKeywordRecommender)))
                 {
                     try
                     {
